Made bulk actoins for whole filtered list work.
